+++ processor/src/main/java/com/linecorp/decaton/processor/runtime/internal/NoopQuotaApplier.java
@@ -1,0 +1,32 @@
+/*
+ * Copyright 2023 LINE Corporation
+ *
+ * LINE Corporation licenses this file to you under the Apache License,
+ * version 2.0 (the "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at:
+ *
+ *   https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
+ * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
+ * License for the specific language governing permissions and limitations
+ * under the License.
+ */
+
+package com.linecorp.decaton.processor.runtime.internal;
+
+import org.apache.kafka.clients.consumer.ConsumerRecord;
+
+import com.linecorp.decaton.processor.runtime.internal.PerKeyQuotaManager.QuotaUsage;
+
+public enum NoopQuotaApplier implements QuotaApplier {
+    INSTANCE;
+
+    @Override
+    public boolean apply(ConsumerRecord<byte[], byte[]> record,
+                         OffsetState offsetState,
+                         QuotaUsage quotaUsage) {
+        return false;
+    }
+}
